+++ Albelli.Assessment.Domain/Models/OrderDto.cs
@@ -15,6 +15,6 @@
         /// <summary>
         /// The minimum bin width.
         /// </summary>
-        public int RequiredBinWidth { get; set; }
+        public decimal RequiredBinWidth { get; set; }
     }
 }
